+++ src/ios/HealthKitManager.swift
@@ -2,52 +2,10 @@ import HealthKit
 
 class HealthKitManager {
     
-    lazy var allVariablesDictToRead: [String: HKObjectType] =
-        [HealthTypeEnum.stepCount.rawValue:HKObjectType.quantityType(forIdentifier: .stepCount)!,
-         HealthTypeEnum.heartRate.rawValue:HKObjectType.quantityType(forIdentifier: .heartRate)!,
-         HealthTypeEnum.bodyMass.rawValue:HKObjectType.quantityType(forIdentifier: .bodyMass)!,
-         HealthTypeEnum.activeEnergyBurned.rawValue:HKObjectType.quantityType(forIdentifier:HKQuantityTypeIdentifier.activeEnergyBurned)!,
-         HealthTypeEnum.height.rawValue:HKObjectType.quantityType(forIdentifier: .height)!]
-    
-    lazy var allVariablesDictToWrite: [String: HKSampleType] =
-        [HealthTypeEnum.stepCount.rawValue:HKSampleType.quantityType(forIdentifier: .stepCount)!,
-         HealthTypeEnum.heartRate.rawValue:HKSampleType.quantityType(forIdentifier: .heartRate)!,
-         HealthTypeEnum.bodyMass.rawValue:HKSampleType.quantityType(forIdentifier: .bodyMass)!,
-         HealthTypeEnum.activeEnergyBurned.rawValue:HKSampleType.quantityType(forIdentifier:HKQuantityTypeIdentifier.activeEnergyBurned)!,
-         HealthTypeEnum.height.rawValue:HKSampleType.quantityType(forIdentifier: .height)!]
-    
-    lazy var profileVariablesDictToRead: [String: HKObjectType] =
-        [HealthTypeEnum.bodyMass.rawValue:HKObjectType.quantityType(forIdentifier: .bodyMass)!,
-         HealthTypeEnum.height.rawValue:HKObjectType.quantityType(forIdentifier: .height)!]
-    
-    lazy var profileVariablesDictToWrite: [String: HKSampleType] =
-        [HealthTypeEnum.stepCount.rawValue:HKSampleType.quantityType(forIdentifier: .stepCount)!,
-         HealthTypeEnum.height.rawValue:HKSampleType.quantityType(forIdentifier: .height)!]
-    
-    lazy var fitnessVariablesDictToRead: [String: HKObjectType] =
-        [HealthTypeEnum.stepCount.rawValue:HKObjectType.quantityType(forIdentifier: .stepCount)!,
-         HealthTypeEnum.activeEnergyBurned.rawValue:HKObjectType.quantityType(forIdentifier:HKQuantityTypeIdentifier.activeEnergyBurned)!]
-    
-    lazy var fitnessVariablesDictToWrite: [String: HKSampleType] =
-        [HealthTypeEnum.stepCount.rawValue:HKSampleType.quantityType(forIdentifier: .stepCount)!,
-         HealthTypeEnum.activeEnergyBurned.rawValue:HKSampleType.quantityType(forIdentifier:HKQuantityTypeIdentifier.activeEnergyBurned)!]
-    
-    lazy var healthVariablesDictToRead: [String: HKObjectType] =
-        [HealthTypeEnum.sleepAnalysis.rawValue:HKSampleType.categoryType(forIdentifier: .sleepAnalysis)!,
-         HealthTypeEnum.heartRate.rawValue:HKObjectType.quantityType(forIdentifier: .heartRate)!]
-    
-    lazy var healthVariablesDictToWrite: [String: HKSampleType] =
-        [HealthTypeEnum.sleepAnalysis.rawValue:HKSampleType.categoryType(forIdentifier: .sleepAnalysis)!,
-         HealthTypeEnum.heartRate.rawValue:HKSampleType.quantityType(forIdentifier: .heartRate)!]
-    
-
     var healthKitTypesToRead = Set<HKObjectType>()
     var healthKitTypesToWrite = Set<HKSampleType>()
-
-    func getData() -> String {
-        return "Test String as result"
-    }
-    
+    var HKTypes = HealthKitTypes()
+        
     func isValidVariable(dict:[String: Any], variable:String) -> Bool {
         let filtered = dict.filter { $0.key == variable }
         return !filtered.isEmpty
@@ -58,17 +16,17 @@ class HealthKitManager {
             for element in permissions {
                 let variable = element.variable
                 
-                let existVariableToRead = isValidVariable(dict: allVariablesDictToRead, variable: variable)
-                let existVariableToWrite = isValidVariable(dict: allVariablesDictToWrite, variable: variable)
+                let existVariableToRead = isValidVariable(dict: HKTypes.allVariablesDictToRead, variable: variable)
+                let existVariableToWrite = isValidVariable(dict: HKTypes.allVariablesDictToWrite, variable: variable)
                 
                 if (!variable.isEmpty) {
                     if (element.accessType == "WRITE" && existVariableToWrite) {
-                        healthKitTypesToWrite.insert(allVariablesDictToWrite[variable]!)
+                        healthKitTypesToWrite.insert(HKTypes.allVariablesDictToWrite[variable]!)
                     }else if (element.accessType == "READWRITE") && existVariableToRead && existVariableToWrite {
-                        healthKitTypesToRead.insert(allVariablesDictToRead[variable]!)
-                        healthKitTypesToWrite.insert(allVariablesDictToWrite[variable]!)
+                        healthKitTypesToRead.insert(HKTypes.allVariablesDictToRead[variable]!)
+                        healthKitTypesToWrite.insert(HKTypes.allVariablesDictToWrite[variable]!)
                     } else if (existVariableToRead) {
-                        healthKitTypesToRead.insert(allVariablesDictToRead[variable]!)
+                        healthKitTypesToRead.insert(HKTypes.allVariablesDictToRead[variable]!)
                     } else {
                         return false
                     }
@@ -84,8 +42,8 @@ class HealthKitManager {
     
     func processVariables(dictToRead:[String: HKObjectType],
                         dictToWrite:[String: HKSampleType],
-                        groupPermissions:GroupPermissions)
-    {
+                        groupPermissions:GroupPermissions) {
+        
         if (groupPermissions.accessType == "WRITE") {
             for item in dictToWrite { healthKitTypesToWrite.insert(item.value) }
         } else if (groupPermissions.accessType == "READWRITE") {
@@ -102,7 +60,7 @@ class HealthKitManager {
                             healthVariables:String,
                             profileVariables:String,
                             summaryVariables:String,
-                            completion: @escaping (Bool, HealthKitAuthorizationErrors?) -> Void) {
+                            completion: @escaping (Bool, NSError?) -> Void) {
         
         var isAuthorizationValid = true
         
@@ -112,48 +70,48 @@ class HealthKitManager {
         
         let all = allVariables.decode(string: allVariables) as GroupPermissions
         if all.isActive {
-            self.processVariables(dictToRead: allVariablesDictToRead,
-                                  dictToWrite: allVariablesDictToWrite,
+            self.processVariables(dictToRead: HKTypes.allVariablesDictToRead,
+                                  dictToWrite: HKTypes.allVariablesDictToWrite,
                                   groupPermissions: all)
         }
         
         let fitness = fitnessVariables.decode(string: fitnessVariables) as GroupPermissions
         if fitness.isActive {
-            self.processVariables(dictToRead: fitnessVariablesDictToRead,
-                                  dictToWrite: fitnessVariablesDictToWrite,
+            self.processVariables(dictToRead: HKTypes.fitnessVariablesDictToRead,
+                                  dictToWrite: HKTypes.fitnessVariablesDictToWrite,
                                   groupPermissions: fitness)
         }
         
         let health = healthVariables.decode(string: healthVariables) as GroupPermissions
         if health.isActive {
-            self.processVariables(dictToRead: healthVariablesDictToRead,
-                                  dictToWrite: healthVariablesDictToWrite,
+            self.processVariables(dictToRead: HKTypes.healthVariablesDictToRead,
+                                  dictToWrite: HKTypes.healthVariablesDictToWrite,
                                   groupPermissions: health)
         }
         
         let profile = profileVariables.decode(string: profileVariables) as GroupPermissions
         if profile.isActive {
-            self.processVariables(dictToRead: profileVariablesDictToRead,
-                                  dictToWrite: profileVariablesDictToWrite,
+            self.processVariables(dictToRead: HKTypes.profileVariablesDictToRead,
+                                  dictToWrite: HKTypes.profileVariablesDictToWrite,
                                   groupPermissions: profile)
         }
         
         let permissonsOK = self.parseCustomPermissons(customPermissions: customPermissions)
         if !permissonsOK {
             isAuthorizationValid = false
-            completion(false, HealthKitAuthorizationErrors.dataTypeNotAvailable)
+            completion(false, HealthKitErrors.dataTypeNotAvailable as NSError)
         }
         
         if (isAuthorizationValid) {
             HKHealthStore().requestAuthorization(toShare: healthKitTypesToWrite,
                                                  read: healthKitTypesToRead) { (success, error) in
                 
-                guard let error = error else {
-                    return completion(false, HealthKitAuthorizationErrors.notAuthorizedByUser)
+                if (error != nil) {
+                    return completion(false, HealthKitErrors.notAuthorizedByUser as NSError)
                 }
                 
                 if success {
-                    completion(success,error as? HealthKitAuthorizationErrors)
+                    completion(success,error as NSError?)
                 }
                 
             }
@@ -161,13 +119,64 @@ class HealthKitManager {
         
     }
     
-    func isHealthDataAvailable() -> HealthKitAuthorizationErrors? {
+    func isHealthDataAvailable() -> NSError? {
+        
         guard HKHealthStore.isHealthDataAvailable() else {
-            return HealthKitAuthorizationErrors.notAvailableOnDevice
+            return HealthKitErrors.notAvailableOnDevice as NSError
         }
         return nil
     }
+    
+    func writeData(variable: String,
+                   value: String,
+                   completion: @escaping (Bool, NSError?) -> Void) {
+        
+        
+        if let error = self.isHealthDataAvailable() {
+            completion(false, error)
+        }
+                
+        guard let type = HKTypes.profileVariablesQuantityDictToWrite[variable] else {
+            let error = HealthKitErrors.dataTypeNotAvailable
+            completion(false, error as NSError)
+            return
+        }
+        
+        guard let unit = HKTypes.profileVariablesUnitDictToWrite[variable] else {
+            let error = HealthKitErrors.dataTypeNotAvailable
+            completion(false, error as NSError)
+            return
+        }
+        
+        if let val = Double(value) {
+            let variableQuantity = HKQuantity(unit: unit, doubleValue: val)
+            
+            let countSample = HKQuantitySample(type: type,
+                                                quantity: variableQuantity,
+                                                   start: Date(),
+                                                   end: Date())
+            
+            HKHealthStore().save(countSample) { (success, error) in
+                
+                if let error = error {
+                    switch (error as NSError).code {
+                    case 1:
+                        completion(false, HealthKitErrors.notAvailableOnDevice as NSError)
+                    case 5:
+                        completion(false, HealthKitErrors.notAuthorizedByUser as NSError)
+                    default:
+                        completion(false, error as NSError)
+                    }
+                    
+                } else {
+                    completion(true, nil)
+                }
+            }
+        }
+       
+    }
 
+    
 }
 
 extension String {