+++ src/ios/OSHealthFitness.swift
@@ -28,25 +28,35 @@ class OSHealthFitness: CordovaImplementation {
                                    summaryVariables:summaryVariables) { [self] (authorized, error) in
             
             if let err = error {
-                self.sendResult(result: "", error:err.localizedDescription , callBackID: self.callbackId)
+                self.sendResult(result: "", error:err , callBackID: self.callbackId)
             }
             
             if authorized {
-                self.sendResult(result: "", error: "", callBackID: self.callbackId)
+                self.sendResult(result: "", error: nil, callBackID: self.callbackId)
             }
         }
     }
     
-    @objc(getData:)
-    func getData(command: CDVInvokedUrlCommand) {
+    @objc(writeData:)
+    func writeData(command: CDVInvokedUrlCommand) {
         callbackId = command.callbackId
-    
-        if let resultStr = plugin?.getData() {
-            self.sendResult(result: resultStr, error: "", callBackID: callbackId)
-        } else {
-            self.sendResult(result: "", error: "Data is empty", callBackID: callbackId)
-        }
         
+        let variable = command.arguments[0] as? String ?? ""
+        let value = command.arguments[1] as? String ?? ""
+        
+        plugin?.writeData(variable: variable, value: value) { success,error in
+            
+            if let err = error {
+                self.sendResult(result: "", error:err, callBackID: self.callbackId)
+            }
+            
+            if success {
+                self.sendResult(result: "", error: nil, callBackID: self.callbackId)
+            }
+            
+        }
+    
     }
+    
     
 }