+++ CHANGELOG.md
@@ -12,6 +12,9 @@ The changes documented here do not include those from the original repository.
 - Implemented WriteProfileData for Android (https://outsystemsrd.atlassian.net/browse/RMET-976)
 
 ## 2021-09-13
+- Implemeted method to write profile data (https://outsystemsrd.atlassian.net/browse/RMET-995)
+
+## 2021-09-13
 - Fixed RequestPermissions for Android (https://outsystemsrd.atlassian.net/browse/RMET-971)
 
 ## 2021-09-10